+++ dpAutoRigSystem/Pipeline/Validator/CheckOut/dpExitEditMode.py
@@ -38,39 +38,42 @@ class ExitEditMode(dpBaseAction.ActionStartClass):
         
         # ---
         # --- validator code --- beginning
-        if objList:
-            toCheckList = objList
-        else:
-            toCheckList = self.dpUIinst.ctrls.getControlList()
-        if toCheckList:
-            self.utils.setProgress(max=len(toCheckList), addOne=False, addNumber=False)
-            for item in toCheckList:
-                self.utils.setProgress(self.dpUIinst.lang[self.title])
-                # conditional to check here
-                if "editMode" in cmds.listAttr(item):
-                    if cmds.getAttr(item+".editMode") == 1:
-                        self.checkedObjList.append(item)
-                        self.foundIssueList.append(True)
-                        if self.firstMode:
-                            self.resultOkList.append(False)
-                        else: #fix
-                            try:
-                                # delete the corrective script job
-                                self.dpUIinst.ctrls.deleteOldJobs(item)
-                                # remove color override
-                                shapeList = cmds.listRelatives(item, shapes=True, children=True, fullPath=True)
-                                if shapeList:
-                                    for shapeNode in shapeList:
-                                        cmds.setAttr(shapeNode+".overrideRGBColors", 0)
-                                # set edit mode off
-                                cmds.setAttr(item+".editMode", 0)
-                                self.resultOkList.append(True)
-                                self.messageList.append(self.dpUIinst.lang['v004_fixed']+": "+item)
-                            except:
+        if not cmds.file(query=True, reference=True):
+            if objList:
+                toCheckList = objList
+            else:
+                toCheckList = self.dpUIinst.ctrls.getControlList()
+            if toCheckList:
+                self.utils.setProgress(max=len(toCheckList), addOne=False, addNumber=False)
+                for item in toCheckList:
+                    self.utils.setProgress(self.dpUIinst.lang[self.title])
+                    # conditional to check here
+                    if "editMode" in cmds.listAttr(item):
+                        if cmds.getAttr(item+".editMode") == 1:
+                            self.checkedObjList.append(item)
+                            self.foundIssueList.append(True)
+                            if self.firstMode:
                                 self.resultOkList.append(False)
-                                self.messageList.append(self.dpUIinst.lang['v005_cantFix']+": "+item)
+                            else: #fix
+                                try:
+                                    # delete the corrective script job
+                                    self.dpUIinst.ctrls.deleteOldJobs(item)
+                                    # remove color override
+                                    shapeList = cmds.listRelatives(item, shapes=True, children=True, fullPath=True)
+                                    if shapeList:
+                                        for shapeNode in shapeList:
+                                            cmds.setAttr(shapeNode+".overrideRGBColors", 0)
+                                    # set edit mode off
+                                    cmds.setAttr(item+".editMode", 0)
+                                    self.resultOkList.append(True)
+                                    self.messageList.append(self.dpUIinst.lang['v004_fixed']+": "+item)
+                                except:
+                                    self.resultOkList.append(False)
+                                    self.messageList.append(self.dpUIinst.lang['v005_cantFix']+": "+item)
+            else:
+                self.notFoundNodes()
         else:
-            self.notFoundNodes()
+            self.notWorkedWellIO(self.dpUIinst.lang['r072_noReferenceAllowed'])
         # --- validator code --- end
         # ---
